+++ demo-app/src/testsAreDemo/SelectedItemsTestsAreDemo.tsx
@@ -17,40 +17,40 @@ export class SelectedItemsTestsAreDemo {
     @ScenarioOptions({linkWithNextScenario: true}) 
     async whenClickOnASegment() {
         // WHEN left click, THEN element is selected
-        await tad.userEventWaitable.click(tad.screenCapturing.getByTestId(testids.item + "_2"));
+        await tad.userEventWaitable.click(tad.screenCapturing.getByTestId('r9k1_' + testids.item + "_2"));
         await this.assertOnlyExpectedSegmentsAreSelected([2]);
 
         //=======HIDDEN TESTS (Not interesting for the user)==========
         tad.demoForEndUserHide();
 
         // WHEN left click again on same element, THEN the same element is selected
-        await tad.userEventWaitable.click(tad.screenCapturing.getByTestId(testids.item + "_2"));
+        await tad.userEventWaitable.click(tad.screenCapturing.getByTestId('r9k1_' + testids.item + "_2"));
         await this.assertOnlyExpectedSegmentsAreSelected([2], true);
         
         // ==Right click can be used also for selecting segments. It works the same as left click (except that right click on a selected segment doesn't change the selection)== 
         // GIVEN nothing is selected
-        await tad.userEventWaitable.click(tad.screenCapturing.getByTestId(testids.row + "_1"));
+        await tad.userEventWaitable.click(tad.screenCapturing.getByTestId('r9k1_' + testids.row + "_1"));
 
         // WHEN right click, THEN element is selected
-        rightClick(tad.screenCapturing.getByTestId(testids.item + "_4"));
+        rightClick(tad.screenCapturing.getByTestId('r9k1_' + testids.item + "_4"));
         await this.assertOnlyExpectedSegmentsAreSelected([4], true);
 
         // WHEN right click again on another element, THEN the new element is selected
-        rightClick(tad.screenCapturing.getByTestId(testids.item + "_5"));
+        rightClick(tad.screenCapturing.getByTestId('r9k1_' + testids.item + "_5"));
         await this.assertOnlyExpectedSegmentsAreSelected([5], true);
 
         // WHEN right click again on same element, THEN the same element is selected
-        rightClick(tad.screenCapturing.getByTestId(testids.item + "_5"));
+        rightClick(tad.screenCapturing.getByTestId('r9k1_' + testids.item + "_5"));
         await this.assertOnlyExpectedSegmentsAreSelected([5], true);
 
         // GIVEN many segments are selected, WHEN right click one of them THEN the selection doesn't change
-        await tad.userEventWaitable.click(tad.screenCapturing.getByTestId(testids.item + "_1"));
-        await tad.userEventWaitable.click(tad.screenCapturing.getByTestId(testids.item + "_2"), { ctrlKey: true });
-        rightClick(tad.screenCapturing.getByTestId(testids.item + "_1"));
+        await tad.userEventWaitable.click(tad.screenCapturing.getByTestId('r9k1_' + testids.item + "_1"));
+        await tad.userEventWaitable.click(tad.screenCapturing.getByTestId('r9k1_' + testids.item + "_2"), { ctrlKey: true });
+        rightClick(tad.screenCapturing.getByTestId('r9k1_' + testids.item + "_1"));
         await this.assertOnlyExpectedSegmentsAreSelected([1, 2], true);
         
         // restore selection for the next scenario
-        await tad.userEventWaitable.click(tad.screenCapturing.getByTestId(testids.item + "_2"));
+        await tad.userEventWaitable.click(tad.screenCapturing.getByTestId('r9k1_' + testids.item + "_2"));
         tad.demoForEndUserShow();
     }
 
@@ -59,7 +59,7 @@ export class SelectedItemsTestsAreDemo {
     async whenClickAnotherSegment() {
         tad.cc(undefined);
          // WHEN left click again on another element, THEN the new element is selected
-         await tad.userEventWaitable.click(tad.screenCapturing.getByTestId(testids.item + "_3"));
+         await tad.userEventWaitable.click(tad.screenCapturing.getByTestId('r9k1_' + testids.item + "_3"));
          await this.assertOnlyExpectedSegmentsAreSelected([3]);
     }
 
@@ -68,39 +68,39 @@ export class SelectedItemsTestsAreDemo {
     async whenCTRLClickOnASegment() {
         // left click + CTRL on another element, THEN the new element is added to selection
         tad.cc("With CTRL Key pressed");
-        await tad.userEventWaitable.click(tad.screenCapturing.getByTestId(testids.item + "_0"), { ctrlKey: true });
+        await tad.userEventWaitable.click(tad.screenCapturing.getByTestId('r9k1_' + testids.item + "_0"), { ctrlKey: true });
         await this.assertOnlyExpectedSegmentsAreSelected([0, 3]);
         
         //=======HIDDEN TESTS (Not interesting for the user)==========
         tad.demoForEndUserHide();
 
         // WHEN left click + SHIFT on another element, THEN the new element is added to selection
-        await tad.userEventWaitable.click(tad.screenCapturing.getByTestId(testids.item + "_4"), { shiftKey: true });
+        await tad.userEventWaitable.click(tad.screenCapturing.getByTestId('r9k1_' + testids.item + "_4"), { shiftKey: true });
         await this.assertOnlyExpectedSegmentsAreSelected([0, 3, 4], true);
 
         // WHEN left click + SHIFT on same element, THEN the element is removed from selection
-        await tad.userEventWaitable.click(tad.screenCapturing.getByTestId(testids.item + "_4"), { shiftKey: true });
+        await tad.userEventWaitable.click(tad.screenCapturing.getByTestId('r9k1_' + testids.item + "_4"), { shiftKey: true });
         await this.assertOnlyExpectedSegmentsAreSelected([0, 3], true);
 
         // WHEN right click + CTRL on another element, THEN the new element is added to selection
-        rightClick(tad.screenCapturing.getByTestId(testids.item + "_4"), { ctrlKey: true });
+        rightClick(tad.screenCapturing.getByTestId('r9k1_' + testids.item + "_4"), { ctrlKey: true });
         await this.assertOnlyExpectedSegmentsAreSelected([0, 3, 4], true);
 
         // WHEN right click + CTRL on same element, THEN the selection doesn't change
-        rightClick(tad.screenCapturing.getByTestId(testids.item + "_4"), { ctrlKey: true });
+        rightClick(tad.screenCapturing.getByTestId('r9k1_' + testids.item + "_4"), { ctrlKey: true });
         await this.assertOnlyExpectedSegmentsAreSelected([0, 3, 4], true);
 
         // WHEN right click + SHIFT on another element, THEN the new element is added to selection
-        rightClick(tad.screenCapturing.getByTestId(testids.item + "_5"), { shiftKey: true });
+        rightClick(tad.screenCapturing.getByTestId('r9k1_' + testids.item + "_5"), { shiftKey: true });
         await this.assertOnlyExpectedSegmentsAreSelected([0, 3, 4, 5], true);
 
         // WHEN right click + SHIFT on same element, THEN the selection doesn't change
-        rightClick(tad.screenCapturing.getByTestId(testids.item + "_5"), { shiftKey: true });
+        rightClick(tad.screenCapturing.getByTestId('r9k1_' + testids.item + "_5"), { shiftKey: true });
         await this.assertOnlyExpectedSegmentsAreSelected([0, 3, 4, 5], true);
 
         // restore selection for the next scenario
-        await tad.userEventWaitable.click(tad.screenCapturing.getByTestId(testids.item + "_0"));
-        await tad.userEventWaitable.click(tad.screenCapturing.getByTestId(testids.item + "_3"), { ctrlKey: true });
+        await tad.userEventWaitable.click(tad.screenCapturing.getByTestId('r9k1_' + testids.item + "_0"));
+        await tad.userEventWaitable.click(tad.screenCapturing.getByTestId('r9k1_' + testids.item + "_3"), { ctrlKey: true });
         tad.demoForEndUserShow();
     }
 
@@ -109,7 +109,7 @@ export class SelectedItemsTestsAreDemo {
     async whenCTRLClickOnASelectedSegment() {
         // WHEN left click + CTRL on same element, THEN the element is removed from selection
         tad.cc("With CTRL Key pressed");
-        await tad.userEventWaitable.click(tad.screenCapturing.getByTestId(testids.item + "_3"), { ctrlKey: true });
+        await tad.userEventWaitable.click(tad.screenCapturing.getByTestId('r9k1_' + testids.item + "_3"), { ctrlKey: true });
         await this.assertOnlyExpectedSegmentsAreSelected([0]);
     }
 
@@ -118,22 +118,22 @@ export class SelectedItemsTestsAreDemo {
     async whenCTRLClickOutside() {
         // When I click outside + Ctrl key, THEN selection doesn't change
         tad.cc("Click outside with CTRL Key pressed");
-        await tad.userEventWaitable.click(tad.screenCapturing.getByTestId(testids.row + "_1"), { ctrlKey: true });
+        await tad.userEventWaitable.click(tad.screenCapturing.getByTestId('r9k1_' + testids.row + "_1"), { ctrlKey: true });
         await this.assertOnlyExpectedSegmentsAreSelected([0]);
 
         //=======HIDDEN TESTS (Not interesting for the user)==========
         // Same as CLICK + CTRL outside any segment works: CLICK + SHIFT, RIGHT CLICK + CTRL, RIGHT CLICK + SHIFT outside any segment
         await tad.demoForEndUserHide()
         // WHEN I click outside + Shift key, THEN selection doesn't change
-        await tad.userEventWaitable.click(tad.screenCapturing.getByTestId(testids.row + "_1"), { shiftKey: true });
+        await tad.userEventWaitable.click(tad.screenCapturing.getByTestId('r9k1_' + testids.row + "_1"), { shiftKey: true });
         await this.assertOnlyExpectedSegmentsAreSelected([0], true);
 
         // WHEN I right click outside + Ctrl key, THEN selection doesn't change
-        rightClick(tad.screenCapturing.getByTestId(testids.row + "_1"), { ctrlKey: true });
+        rightClick(tad.screenCapturing.getByTestId('r9k1_' + testids.row + "_1"), { ctrlKey: true });
         await this.assertOnlyExpectedSegmentsAreSelected([0], true);
 
         // WHEN I click outside + Shift key, THEN selection doesn't change
-        rightClick(tad.screenCapturing.getByTestId(testids.row + "_1"), { shiftKey: true });
+        rightClick(tad.screenCapturing.getByTestId('r9k1_' + testids.row + "_1"), { shiftKey: true });
         await this.assertOnlyExpectedSegmentsAreSelected([0], true);
         tad.demoForEndUserShow();
     }
@@ -143,16 +143,16 @@ export class SelectedItemsTestsAreDemo {
     async whenClickOutside() {
          // WHEN left click on row, THEN no segment is selected
          tad.cc("Left click outside any segment");
-         await tad.userEventWaitable.click(tad.screenCapturing.getByTestId(testids.row + "_1"));
+         await tad.userEventWaitable.click(tad.screenCapturing.getByTestId('r9k1_' + testids.row + "_1"));
          await tad.showSpotlight({ message: "No segment is selected", focusOnLastElementCaptured: false });
          await this.assertOnlyExpectedSegmentsAreSelected([]);
  
          //=======HIDDEN TESTS (Not interesting for the user)==========
          // GIVEN a segment is selected
          tad.demoForEndUserHide();
-         rightClick(tad.screenCapturing.getByTestId(testids.item + "_2"));
+         rightClick(tad.screenCapturing.getByTestId('r9k1_' + testids.item + "_2"));
          // WHEN right click on row, THEN no segment is selected
-         rightClick(tad.screenCapturing.getByTestId(testids.row + "_1"));
+         rightClick(tad.screenCapturing.getByTestId('r9k1_' + testids.row + "_1"));
          await this.assertOnlyExpectedSegmentsAreSelected([], true);
          tad.demoForEndUserShow();
     }
@@ -176,7 +176,7 @@ export class SelectedItemsTestsAreDemo {
         // Using right mouse button works the same
         tad.demoForEndUserHide();
         // GIVEN nothing selected
-        await tad.userEventWaitable.click(tad.screenCapturing.getByTestId(testids.row + "_1"));
+        await tad.userEventWaitable.click(tad.screenCapturing.getByTestId('r9k1_' + testids.row + "_1"));
 
         await this.dragToSelect(0, 1, 3, 3, true);
         await this.assertOnlyExpectedSegmentsAreSelected([0, 3], true);
@@ -206,7 +206,7 @@ export class SelectedItemsTestsAreDemo {
         // Same as LEFT button + CTRL works: LEFT + SHIFT, RIGHT + CTRL, RIGHT + SHIFT
         tad.demoForEndUserHide();
         // unselect everything
-        await tad.userEventWaitable.click(tad.screenCapturing.getByTestId(testids.row + "_1"));
+        await tad.userEventWaitable.click(tad.screenCapturing.getByTestId('r9k1_' + testids.row + "_1"));
          // WHEN I draw a rectangle containing segments 0, 3 with LEFT mouse button and SHIFT pressed, THEN segments 0, 3 are added to selection
         await this.dragToSelect(0, 1, 3, 3, true, true);
         await this.assertOnlyExpectedSegmentsAreSelected([0, 3], true);
@@ -215,7 +215,7 @@ export class SelectedItemsTestsAreDemo {
         await this.assertOnlyExpectedSegmentsAreSelected([0], true);
 
         // unselect everything
-        await tad.userEventWaitable.click(tad.screenCapturing.getByTestId(testids.row + "_1"));
+        await tad.userEventWaitable.click(tad.screenCapturing.getByTestId('r9k1_' + testids.row + "_1"));
         // WHEN I draw a rectangle containing segments 0, 3 with LEFT mouse button and SHIFT pressed, THEN segments 0, 3 are selected
         await this.dragToSelect(0, 1, 3, 3, false, false, true);
         await this.assertOnlyExpectedSegmentsAreSelected([0, 3], true);
@@ -224,7 +224,7 @@ export class SelectedItemsTestsAreDemo {
         await this.assertOnlyExpectedSegmentsAreSelected([0], true);
 
         // unselect everything
-        await tad.userEventWaitable.click(tad.screenCapturing.getByTestId(testids.row + "_1"));
+        await tad.userEventWaitable.click(tad.screenCapturing.getByTestId('r9k1_' + testids.row + "_1"));
         // WHEN I draw a rectangle containing segments 0, 3 with RIGHT mouse button and SHIFT pressed, THEN segments 0, 3 are selected
         await this.dragToSelect(0, 1, 3, 3, true, false, true);
         await this.assertOnlyExpectedSegmentsAreSelected([0, 3], true);
@@ -233,7 +233,7 @@ export class SelectedItemsTestsAreDemo {
         await this.assertOnlyExpectedSegmentsAreSelected([0], true);
 
         // Select again the segment 0
-        await tad.userEventWaitable.click(tad.screenCapturing.getByTestId(testids.row + "_1"));
+        await tad.userEventWaitable.click(tad.screenCapturing.getByTestId('r9k1_' + testids.row + "_1"));
         await this.dragToSelect(0, 0, 0, 0, false, true);
 
         tad.demoForEndUserShow();
@@ -244,16 +244,16 @@ export class SelectedItemsTestsAreDemo {
     async whenCTRLDragASelectionRectangleOverAnEmptyArea() {
         tad.cc(undefined);
         await tad.showSpotlight({ message: "WHEN hold CTRL + drag a selection rectangle over an empty area, THEN the selection doesn't change", focusOnLastElementCaptured: false });
-        let startingRow = tad.screenCapturing.getByTestId(testids.row + "_0");
-        tad.getObjectViaCheat(Timeline).dragStart(startingRow, 5);
-        await tad.getObjectViaCheat(Timeline).dragMove(10, 10, 5);
-        tad.getObjectViaCheat(Timeline).dragEnd({ctrlKey: true});
+        let startingRow = tad.screenCapturing.getByTestId('r9k1_' + testids.row + "_0");
+        tad.getObjectViaCheat(Timeline, 'r9k1').dragStart(startingRow, 5);
+        await tad.getObjectViaCheat(Timeline, 'r9k1').dragMove(10, 10, 5);
+        tad.getObjectViaCheat(Timeline, 'r9k1').dragEnd({ctrlKey: true});
         await this.assertOnlyExpectedSegmentsAreSelected([0]);
         
         //=======HIDDEN TESTS (Not interesting for the user)==========
         // Using right mouse button + CTRL works the same
         tad.demoForEndUserHide();
-        let row = tad.screenCapturing.getByTestId(testids.row + "_0");
+        let row = tad.screenCapturing.getByTestId('r9k1_' + testids.row + "_0");
         await tad.fireEventWaitable.mouseDown(row, { clientX: row.getBoundingClientRect().x + 5, clientY: row.getBoundingClientRect().y + 5, button: 2, ctrlKey: true });
         await tad.fireEventWaitable.mouseMove(row, { clientX: 10, clientY: 10, pageX: 10, ctrlKey: true });
         await tad.fireEventWaitable.mouseUp(row, { button: 2, ctrlKey: true });
@@ -266,10 +266,10 @@ export class SelectedItemsTestsAreDemo {
     async whenDragASelectionRectangleOverAnEmptyArea() {
         tad.cc(undefined);
         await tad.showSpotlight({ message: "I drag to select outside any segments (using left mouse button)", focusOnLastElementCaptured: false });
-        let startingRow = tad.screenCapturing.getByTestId(testids.row + "_0");
-        tad.getObjectViaCheat(Timeline).dragStart(startingRow, 5);
-        await tad.getObjectViaCheat(Timeline).dragMove(10, 10, 5);
-        tad.getObjectViaCheat(Timeline).dragEnd();
+        let startingRow = tad.screenCapturing.getByTestId('r9k1_' + testids.row + "_0");
+        tad.getObjectViaCheat(Timeline, 'r9k1').dragStart(startingRow, 5);
+        await tad.getObjectViaCheat(Timeline, 'r9k1').dragMove(10, 10, 5);
+        tad.getObjectViaCheat(Timeline, 'r9k1').dragEnd();
         await tad.showSpotlight({ message: "No segment is selected", focusOnLastElementCaptured: false });
         await this.assertOnlyExpectedSegmentsAreSelected([]);
         
@@ -277,9 +277,9 @@ export class SelectedItemsTestsAreDemo {
         // Using right mouse button works the same
         tad.demoForEndUserHide();
         // given a selected segment
-        await tad.userEventWaitable.click(tad.screenCapturing.getByTestId(testids.item + "_3"));
+        await tad.userEventWaitable.click(tad.screenCapturing.getByTestId('r9k1_' + testids.item + "_3"));
 
-        let row = tad.screenCapturing.getByTestId(testids.row + "_0");
+        let row = tad.screenCapturing.getByTestId('r9k1_' + testids.row + "_0");
         await tad.fireEventWaitable.mouseDown(row, { clientX: row.getBoundingClientRect().x + 5, clientY: row.getBoundingClientRect().y + 5, button: 2 });
         await tad.fireEventWaitable.mouseMove(row, { clientX: 10, clientY: 10, pageX: 10 });
         await tad.fireEventWaitable.mouseUp(row, { button: 2 });
@@ -292,12 +292,12 @@ export class SelectedItemsTestsAreDemo {
     ////////////////////////////////////////////////////////////////////////////////////////
 
     async dragToSelect(startingRowIndex, endingRowIndex, startingSegmentIndex, endingSegmentIndex, rightClick?, ctrlKey = false, shiftKey = false) {
-        let startingRow = tad.screenCapturing.getByTestId(testids.row + "_" + startingRowIndex);
+        let startingRow = tad.screenCapturing.getByTestId('r9k1_' + testids.row + "_" + startingRowIndex);
         let startingRowRect = startingRow.getBoundingClientRect();
-        let endingRow = tad.screenCapturing.getByTestId(testids.row + "_" + endingRowIndex);
+        let endingRow = tad.screenCapturing.getByTestId('r9k1_' + testids.row + "_" + endingRowIndex);
         let endingRowRect = endingRow.getBoundingClientRect();
-        const startingSegmentRect = tad.screenCapturing.getByTestId(testids.item + "_" + startingSegmentIndex).getBoundingClientRect();
-        const endingSegmentRect = tad.screenCapturing.getByTestId(testids.item + "_" + endingSegmentIndex).getBoundingClientRect();
+        const startingSegmentRect = tad.screenCapturing.getByTestId('r9k1_' + testids.item + "_" + startingSegmentIndex).getBoundingClientRect();
+        const endingSegmentRect = tad.screenCapturing.getByTestId('r9k1_' + testids.item + "_" + endingSegmentIndex).getBoundingClientRect();
         const deltaX = endingSegmentRect.x + endingSegmentRect.width - startingSegmentRect.x;
         const deltaY = endingRowRect.y + endingRowRect.height - startingRowRect.y;
 
@@ -314,15 +314,15 @@ export class SelectedItemsTestsAreDemo {
         } else {
             // 150 is the group offset
             // we needed to subtract -5 because else the selection rectangle (that snapps to row) will get till the endingRow + 1, instead endingRow
-            tad.getObjectViaCheat(Timeline).dragStart(startingRow, startingSegmentRect.x - 150);
-            await tad.getObjectViaCheat(Timeline).dragMove(deltaX, deltaY - 5, 5);
-            tad.getObjectViaCheat(Timeline).dragEnd({ ctrlKey: ctrlKey, shiftKey: shiftKey });
+            tad.getObjectViaCheat(Timeline, 'r9k1').dragStart(startingRow, startingSegmentRect.x - 150);
+            await tad.getObjectViaCheat(Timeline, 'r9k1').dragMove(deltaX, deltaY - 5, 5);
+            tad.getObjectViaCheat(Timeline, 'r9k1').dragEnd({ ctrlKey: ctrlKey, shiftKey: shiftKey });
         }
     }
 
     async assertOnlyExpectedSegmentsAreSelected(expectedSelectedSegments: number[], demoForEndUserHide?) {
         for (var i = 0; i < someTasks.length; i++) {
-            const segment = tad.screenCapturing.getByTestId(testids.item + "_" + i);
+            const segment = tad.screenCapturing.getByTestId('r9k1_' + testids.item + "_" + i);
             if (expectedSelectedSegments.indexOf(i) >= 0) {
                 tad.cc("Segment " + i + " is selected (has resize anchors, brighter color and shadow effect)");
                 // await tad.assertWaitable.include(Array.from(segment.classList), "rct9k-items-outer-selected");