+++ src/components/ContextMenu/IAction.ts
@@ -2,6 +2,9 @@ import React from "react"
 import { IconProps, SemanticShorthandItem } from "semantic-ui-react"
 import { Point } from "./ContextMenu"
 
+// TODO RM34271 let's get rid of IActionParamForRun. It overcomplicates the inheritance. We'll have here closeContextMenu() that won't do anything.
+// export interface IActionParam<S> {
+// selection: S[]
 export interface IActionParam {
     selection: any[]
 }
@@ -22,16 +25,22 @@ export interface IActionParamForRun extends IActionParam {
     eventPoint?: Point
 }
 
+// TODO RM34271 CS: for the moment it's not clear why we need this. It's not used by the lib. And currently, the task of opening
+// the context menu belongs to the user. This interface, judging by its name, seems to be needed for the process of opening the CM
+/**
+ * We may add in the future further customization possibilities. E.g. a `render()` function, etc.
+ */
 export interface IOnContextMenuShowParam {
     actionParam: IActionParam
-    // poate mai bagam pe viitor chestii aici; precum posib de alte customizari; precum o functie de render pentru context menu
 }
 
 /**
  * This is a descriptor passed by the user to define an action entry in the context menu.
  * Can have a `run` function that will be called when user will clicks the corresponding menu entry. 
- * It will receive a parameter containing things like: current selection, a close menu function to be run after clicking the action
+ * It will receive a parameter containing things like: current selection, a close menu function to be run after clicking the action.
  */
+// TODO RM34271 
+// export interface IAction<AP extends IActionParam> {
 export interface IAction {
     /**
      * Should return true of false whether or not the action is visible for the selected items received as parameter